# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mail Gateway API
  version: "1.0.0"
  description: |
    Mail Gateway REST API 스펙입니다. 이 문서는 메일 수신, 분류, 전달 관련 엔드포인트를 정의합니다.
servers:
  - url: http://localhost:3000
    description: 로컬 개발 서버
paths:
...
  /api/emails:
    get:
      tags:
        - Emails
      summary: 이메일 목록 (페이지네이션)
      description: 이메일 목록을 페이지네이션으로 조회합니다.
      parameters:
        - name: status
          in: query
          description: 이메일 상태 필터
          schema:
            type: string
            enum: [received, classified, forwarded, failed, archived]
        - name: tenantId
          in: query
          description: 테넌트 ID 필터
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: 페이지 크기
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 이메일 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmailSummary'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /api/emails/{id}:
    get:
      tags:
        - Emails
      summary: 이메일 상세
      description: 이메일 상세 정보, 첨부파일과 전달 로그를 포함합니다.
      parameters:
        - name: id
          in: path
          required: true
          description: 이메일 ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 이메일 상세
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/EmailDetail'
        '404':
          description: 이메일을 찾을 수 없음
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Email not found
  /api/tenants:
    get:
      tags:
        - Tenants
      summary: 테넌트 목록
      description: 모든 테넌트를 반환합니다.
      responses:
        '200':
          description: 테넌트 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
    post:
      tags:
        - Tenants
      summary: 테넌트 생성
      description: 새 테넌트를 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - domains
              properties:
                name:
                  type: string
                domains:
                  type: string
                  description: >-
                    JSON 배열 문자열로 전달합니다. 예: ["example.com"]
                settings:
                  type: string
                  nullable: true
                  description: JSON 문자열
                enabled:
                  type: boolean
                  default: true
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tenant'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Name is required
...
  /api/rules:
    get:
      tags:
        - Rules
      summary: 분류 룰 목록
      description: 분류 룰 목록을 조회합니다.
      parameters:
        - name: tenantId
          in: query
          description: 테넌트별 필터
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 룰 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClassificationRule'
    post:
      tags:
        - Rules
      summary: 룰 생성
      description: 새 분류 룰을 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, conditions, action]
              properties:
                name:
                  type: string
                conditions:
                  type: string
                  description: JSON 배열 문자열 (RuleCondition[])
                action:
                  type: string
                  enum: [forward, log, archive, reject]
                description:
                  type: string
                  nullable: true
                priority:
                  type: integer
                  default: 0
                enabled:
                  type: boolean
                  default: true
                forwardTo:
                  type: string
                  nullable: true
                category:
                  type: string
                  nullable: true
                tenantId:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ClassificationRule'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /api/forward-logs:
    get:
      tags:
        - ForwardLogs
      summary: 전달 로그 목록 (페이지네이션)
      description: 전달 로그를 페이지네이션으로 조회합니다.
      parameters:
        - name: emailId
          in: query
          description: 이메일 ID 필터
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: 전달 상태 필터
          schema:
            type: string
            enum: [pending, success, failed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 전달 로그 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ForwardLog'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
...
components:
  schemas:
    EmailSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        mailFrom:
          type: string
        rcptTo:
          type: string
        remoteIp:
          type: string
        messageId:
          type: string
          nullable: true
        subject:
          type: string
          nullable: true
        fromHeader:
          type: string
          nullable: true
        toHeader:
          type: string
          nullable: true
        ccHeader:
          type: string
          nullable: true
        replyTo:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
          nullable: true
        headers:
          type: string
          nullable: true
          description: JSON 문자열
        textBody:
          type: string
          nullable: true
        htmlBody:
          type: string
          nullable: true
        status:
          type: string
          enum: [received, classified, forwarded, failed, archived]
        category:
          type: string
          nullable: true
        matchedRule:
          type: string
          nullable: true
        tenantId:
          type: string
          format: uuid
          nullable: true
        _count:
          type: object
          properties:
            attachments:
              type: integer
            forwardLogs:
              type: integer
    EmailDetail:
      allOf:
        - $ref: '#/components/schemas/EmailSummary'
        - type: object
          properties:
            rawMessage:
              type: string
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
            forwardLogs:
              type: array
              items:
                $ref: '#/components/schemas/ForwardLog'
    Attachment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        emailId:
          type: string
          format: uuid
        filename:
          type: string
          nullable: true
        contentType:
          type: string
        size:
          type: integer
        checksum:
          type: string
          nullable: true
        storagePath:
          type: string
          nullable: true
    ForwardLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        emailId:
          type: string
          format: uuid
        forwardTo:
          type: string
        status:
          type: string
          enum: [pending, success, failed]
        error:
          type: string
          nullable: true
        smtpResponse:
          type: string
          nullable: true
        attempts:
          type: integer
          default: 1
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          type: string
        domains:
          type: array
          items:
            type: string
          description: 도메인 배열 (기본적으로 DB에는 JSON으로 저장)
        settings:
          type: string
          nullable: true
          description: JSON 문자열
        enabled:
          type: boolean
    ClassificationRule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        name:
          type: string
        description:
          type: string
          nullable: true
        priority:
          type: integer
          default: 0
        enabled:
          type: boolean
          default: true
        conditions:
          type: string
          description: JSON 배열 문자열 (RuleCondition[])
        action:
          type: string
          enum: [forward, log, archive, reject]
        forwardTo:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        tenantId:
          type: string
          format: uuid
          nullable: true
...
security: []
...